CoremlBugFixes:Input variable name can be something other than data; … (#7746)
* CoremlBugFixes:Input variable name can be something other than data; making pad and stride optional.

Earlier, we were not providing data_names argument while creating the module which meant that the input data variable name was assumed to be "data". This is fixed. Also, added a unit test for it (due to which utils.load_model(..) had to be refactored).

The second bug was we missed assuming pad and stride parameters for convolutional layers are optional arguments. Added a unit test for this too.

Also, tested with mnist model from the tutorial (by changing the input variable name to something other than data)

* Minor rewording of a unit test.