Fix missing of CUDA device on non-GPU host (#7855)
* Fix missing of CUDA device on non-GPU host

The issue that "no CUDA-capable device is detected" occurs when calling MxNet from Java/Scala on a host without GPU. 

The user scenario is to run MxNet model on CPU-only host with Context input as cpu(0). The default GPU unit shares the same index of zero with the default CPU unit. By default the MXNET_USE_CUDA is enabled and the logic will proceed to the line of "CUDA_CALL". The calling of cudaSetDevice exits with exception since there is GPU found for index of zero. The proposal here is to enable the "CUDA_CALL" only when the device type of Context is not CPU.

* Revise conditional statement to be more readable

Revise the conditional statement about when the cudaSetDevice could be called to be more readable.

* Re-run the CI test

Trivial change to re-run the continuous integration test.

* Lines should be equal or less than 100 characters

Lines should be equal or less than 100 characters

* Fix line ending in whitespace

Fix line ending in whitespace